Initial version!
I admittedly took a while to actually create this initial version.  For
the first 2 months after creating the repository, I did nothing, but
here it is, the initial version of the game!

It uses TinyXML2 to parse XML data and std library for common data
types, like strings.  I also included a nifty addition operator overload
function for making string concatenation easier.  This game was made
using the CodeBlocks IDE but should compile easily on other IDEs as
well.